+++ modules/tinymce/src/themes/silver/test/ts/browser/statusbar/StatusbarTest.ts
@@ -87,6 +87,7 @@ describe('browser.tinymce.themes.silver.statusbar.StatusbarTest', () => {
           children: [
             s.anything(),
             s.anything(),
+            s.anything(),
             s.element('div', {
               classes: [ arr.has('tox-statusbar') ],
               children: [
@@ -128,6 +129,7 @@ describe('browser.tinymce.themes.silver.statusbar.StatusbarTest', () => {
           children: [
             s.anything(),
             s.anything(),
+            s.anything(),
             s.element('div', {
               classes: [ arr.has('tox-statusbar') ],
               children: fullStatusbarSpec(s, str, arr)
@@ -150,6 +152,7 @@ describe('browser.tinymce.themes.silver.statusbar.StatusbarTest', () => {
           children: [
             s.anything(),
             s.anything(),
+            s.anything(),
             s.element('div', {
               classes: [ arr.has('tox-statusbar') ],
               children: statusbarWithoutWordcountSpec(s, str, arr)
@@ -170,6 +173,7 @@ describe('browser.tinymce.themes.silver.statusbar.StatusbarTest', () => {
         s.element('div', {
           classes: [ arr.has('tox-editor-container') ],
           children: [
+            s.anything(),
             s.anything(),
             s.anything(),
             s.element('div', {
@@ -223,6 +227,7 @@ describe('browser.tinymce.themes.silver.statusbar.StatusbarTest', () => {
             children: [
               s.anything(),
               s.anything(),
+              s.anything(),
               s.element('div', {
                 classes: [ arr.has('tox-statusbar') ],
                 children: [
@@ -259,6 +264,7 @@ describe('browser.tinymce.themes.silver.statusbar.StatusbarTest', () => {
           s.element('div', {
             classes: [ arr.has('tox-editor-container') ],
             children: [
+              s.anything(),
               s.anything(),
               s.anything(),
               s.element('div', {